custom middleware for ethers-rs - 1 error on get_balance
